+++ Flux/Packages/FluxCore/Sources/FluxCore/Widget/WidgetKinds.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum WidgetKinds {
+    public static let battery = "me.nore.ig.flux.widget.battery"
+    public static let accessory = "me.nore.ig.flux.widget.accessory"
+}